int tests fixed

+++ integration/test/010_projects.Tests.ps1
@@ -13,10 +13,49 @@ Describe 'VSTeam Integration Tests' {
       $acct = $env:ACCT
       $api = $env:API_VERSION
       $email = $env:EMAIL
+      $projectName = 'TeamModuleIntegration' + [guid]::NewGuid().toString().substring(0, 5)
+      $newProjectName = $projectName + [guid]::NewGuid().toString().substring(0, 5) + '1'
+
+      $originalLocation = Get-Location
+
+      $profile = Get-Profile | Where-Object url -like "*$acct*"
+
+      if ($profile) {
+         # Save current profile data
+         $oAcct = $profile.URL
+         $oVersion = $profile.Version
+         $oName = $profile.Name
+      }
 
       Add-VSTeamProfile -Account $acct -PersonalAccessToken $pat -Version $api -Name intTests
       Add-VSTeamAccount -Profile intTests -Drive int
-   }   
+   }
+
+   Context 'Set-VSTeamDefaultProject' {
+      It 'should set default project' {
+         Set-VSTeamDefaultProject 'MyProject'
+
+         $Global:PSDefaultParameterValues['*:projectName'] | Should be 'MyProject'
+      }
+
+      It 'should update default project' {
+         $Global:PSDefaultParameterValues['*:projectName'] = 'MyProject'
+
+         Set-VSTeamDefaultProject -Project 'NextProject'
+
+         $Global:PSDefaultParameterValues['*:projectName'] | Should be 'NextProject'
+      }
+   }
+
+   Context 'Clear-VSTeamDefaultProject' {
+      It 'should clear default project' {
+         $Global:PSDefaultParameterValues['*:projectName'] = 'MyProject'
+
+         Clear-VSTeamDefaultProject
+
+         $Global:PSDefaultParameterValues['*:projectName'] | Should BeNullOrEmpty
+      }
+   }
 
    Context 'Profile full exercise' {
       It 'Get-VSTeamProfile' {
@@ -30,23 +69,23 @@ Describe 'VSTeam Integration Tests' {
 
    Context 'Project full exercise' {
       It 'Add-VSTeamProject Should create project' {
-         Add-VSTeamProject -ProjectName 'TeamModuleIntegration' | Should Not Be $null
+         Add-VSTeamProject -ProjectName $projectName | Should Not Be $null
       }
 
       It 'Get-VSTeamProject Should return projects' {
-         Get-VSTeamProject -ProjectName 'TeamModuleIntegration'  | Should Not Be $null
+         Get-VSTeamProject -ProjectName $projectName  | Should Not Be $null
       }
 
       It 'Update-VSTeamProject Should update description' {
-         Update-VSTeamProject -ProjectName 'TeamModuleIntegration' -NewDescription 'Test Description' -Force
+         Update-VSTeamProject -ProjectName $projectName -NewDescription 'Test Description' -Force
 
-         Get-VSTeamProject -ProjectName 'TeamModuleIntegration' | Select-Object -ExpandProperty 'Description' | Should Be 'Test Description'
+         Get-VSTeamProject -ProjectName $projectName | Select-Object -ExpandProperty 'Description' | Should Be 'Test Description'
       }
 
       It 'Update-VSTeamProject Should update name' {
-         Update-VSTeamProject -ProjectName 'TeamModuleIntegration' -NewName 'TeamModuleIntegration1' -Force
+         Update-VSTeamProject -ProjectName $projectName -NewName $newProjectName -Force
 
-         Get-VSTeamProject -ProjectName 'TeamModuleIntegration1' | Select-Object -ExpandProperty 'Description' | Should Be 'Test Description'
+         Get-VSTeamProject -ProjectName $newProjectName | Select-Object -ExpandProperty 'Description' | Should Be 'Test Description'
       }
    }
 
@@ -63,7 +102,7 @@ Describe 'VSTeam Integration Tests' {
       }
 
       It 'Should list Builds, Releases and Teams' {
-         Set-Location 'TeamModuleIntegration1'
+         Set-Location $newProjectName
          Get-ChildItem | Should Not Be $null
       }
 
@@ -75,18 +114,18 @@ Describe 'VSTeam Integration Tests' {
    
    Context 'Git full exercise' {
       It 'Get-VSTeamGitRepository Should return repository' {
-         Get-VSTeamGitRepository -ProjectName 'TeamModuleIntegration1' | Select-Object -ExpandProperty Name | Should Be 'TeamModuleIntegration1'
+         Get-VSTeamGitRepository -ProjectName $newProjectName | Select-Object -ExpandProperty Name | Should Be $newProjectName
       }
 
       It 'Add-VSTeamGitRepository Should create repository' {
-         Add-VSTeamGitRepository -ProjectName 'TeamModuleIntegration1' -Name 'testing'
-         (Get-VSTeamGitRepository -ProjectName 'TeamModuleIntegration1').Count | Should Be 2
-         Get-VSTeamGitRepository -ProjectName 'TeamModuleIntegration1' -Name 'testing' | Select-Object -ExpandProperty Name | Should Be 'testing'
+         Add-VSTeamGitRepository -ProjectName $newProjectName -Name 'testing'
+         (Get-VSTeamGitRepository -ProjectName $newProjectName).Count | Should Be 2
+         Get-VSTeamGitRepository -ProjectName $newProjectName -Name 'testing' | Select-Object -ExpandProperty Name | Should Be 'testing'
       }
 
       It 'Remove-VSTeamGitRepository Should delete repository' {
-         Get-VSTeamGitRepository -ProjectName 'TeamModuleIntegration1' -Name 'testing' | Select-Object -ExpandProperty Id | Remove-VSTeamGitRepository -Force
-         Get-VSTeamGitRepository -ProjectName 'TeamModuleIntegration1' | Where-Object { $_.Name -eq 'testing' } | Should Be $null
+         Get-VSTeamGitRepository -ProjectName $newProjectName -Name 'testing' | Select-Object -ExpandProperty Id | Remove-VSTeamGitRepository -Force
+         Get-VSTeamGitRepository -ProjectName $newProjectName | Where-Object { $_.Name -eq 'testing' } | Should Be $null
       }
    }
 
@@ -104,23 +143,21 @@ Describe 'VSTeam Integration Tests' {
    }
 
    Context 'Queue full exercise' {
-      $id = 0
-
       It 'Get-VSTeamQueue Should return agent Queues' {
-         $actual = Get-VSTeamQueue -ProjectName 'TeamModuleIntegration1'
+         $actual = Get-VSTeamQueue -ProjectName $newProjectName
 
          if ($acct -like "http://*") {
+            $global:queueId = $actual.id
             $actual.name | Should Be 'Default'
-            $id = $actual.id
          }
          else {
+            $global:queueId = $actual[0].id
             $actual.Count | Should Be 5
-            $id = $actual[0].id
          }
       }
 
       It 'Get-VSTeamQueue By Id Should return agent Queue' {
-         $actual = Get-VSTeamQueue -ProjectName 'TeamModuleIntegration1' -Id $id
+         $actual = Get-VSTeamQueue -ProjectName $newProjectName -Id $global:queueId
          
          $actual | Should Not Be $null
       }
@@ -128,83 +165,86 @@ Describe 'VSTeam Integration Tests' {
 
    Context 'Service Endpoints full exercise' {
       It 'Add-VSTeamSonarQubeEndpoint Should add servcie endpoint' {
-         Add-VSTeamSonarQubeEndpoint -ProjectName 'TeamModuleIntegration1' -EndpointName 'TestSonarQube' `
+         Add-VSTeamSonarQubeEndpoint -ProjectName $newProjectName -EndpointName 'TestSonarQube' `
             -SonarQubeURl 'http://sonarqube.somewhereIn.cloudapp.azure.com:9000' -PersonalAccessToken 'Faketoken' | Should Not Be $null
       }
 
       It 'Add-VSTeamAzureRMServiceEndpoint Should add servcie endpoint' {
-         Add-VSTeamAzureRMServiceEndpoint -ProjectName 'TeamModuleIntegration1' -displayName 'AzureEndoint' `
+         Add-VSTeamAzureRMServiceEndpoint -ProjectName $newProjectName -displayName 'AzureEndoint' `
             -subscriptionId '00000000-0000-0000-0000-000000000000' -subscriptionTenantId '00000000-0000-0000-0000-000000000000' `
             -servicePrincipalId '00000000-0000-0000-0000-000000000000' -servicePrincipalKey 'fakekey' | Should Not Be $null
       }
 
       It 'Get-VSTeamServiceEndpoint Should return service endpoints' {
-         $actual = Get-VSTeamServiceEndpoint -ProjectName 'TeamModuleIntegration1'
+         $actual = Get-VSTeamServiceEndpoint -ProjectName $newProjectName
 
          $actual.Count | Should Be 2
       }
 
       It 'Remove-VSTeamServiceEndpoint Should delete service endpoints' {
-         Get-VSTeamServiceEndpoint -ProjectName 'TeamModuleIntegration1' | Remove-VSTeamServiceEndpoint -ProjectName 'TeamModuleIntegration1' -Force
+         Get-VSTeamServiceEndpoint -ProjectName $newProjectName | Remove-VSTeamServiceEndpoint -ProjectName $newProjectName -Force
 
-         Get-VSTeamServiceEndpoint -ProjectName 'TeamModuleIntegration1' | Should Be $null
+         Get-VSTeamServiceEndpoint -ProjectName $newProjectName | Should Be $null
       }
    }
 
-   Context 'Users exercise' {
+   # Not supported on TFS
+   if (-not ($acct -like "http://*")) {
+      Context 'Users exercise' {
     
-      It 'Get-VSTeamUser Should return all usrs' {
-         Get-VSTeamUser | Should Not Be $null         
-      }
+         It 'Get-VSTeamUser Should return all usrs' {
+            Get-VSTeamUser | Should Not Be $null         
+         }
 
-      It 'Get-VSTeamUser ById Should return Teams' {
-         $id = (Get-VSTeamUser | Where-Object email -eq $email).Id
-         Get-VSTeamUser -Id $id | Should Not Be $null
-      }
+         It 'Get-VSTeamUser ById Should return Teams' {
+            $id = (Get-VSTeamUser | Where-Object email -eq $email).Id
+            Get-VSTeamUser -Id $id | Should Not Be $null
+         }
 
-      It 'Remove-VSTeamUser should fail' {
-         { Remove-VSTeamUser -Email fake@NoteReal.foo -Force } | Should Throw
-      }
+         It 'Remove-VSTeamUser should fail' {
+            { Remove-VSTeamUser -Email fake@NoteReal.foo -Force } | Should Throw
+         }
       
-      It 'Remove-VSTeamUser should delete the team' {
-         Remove-VSTeamUser -Email $email -Force
-         Get-VSTeamUser | Where-Object Email -eq $email | Should Be $null
-      }
+         It 'Remove-VSTeamUser should delete the team' {
+            Remove-VSTeamUser -Email $email -Force
+            Get-VSTeamUser | Where-Object Email -eq $email | Should Be $null
+         }
 
-      It 'Add-VSTeamUser should add a team' {
-         Add-VSTeamUser -Email $email | Should Not Be $null
-         (Get-VSTeamUser).Count | Should Be 2
+         It 'Add-VSTeamUser should add a team' {
+            Add-VSTeamUser -Email $email -License StakeHolder | Should Not Be $null
+            (Get-VSTeamUser).Count | Should Be 2
+         }
       }
    }
 
    Context 'Teams full exercise' {
       It 'Get-VSTeam ByName Should return Teams' {
-         Get-VSTeam -ProjectName 'TeamModuleIntegration1' -Name 'TeamModuleIntegration1 Team' | Should Not Be $null
+         Get-VSTeam -ProjectName $newProjectName -Name "$newProjectName Team" | Should Not Be $null
       }
 
-      $id = (Get-VSTeam -ProjectName 'TeamModuleIntegration1').Id
+      $global:id = (Get-VSTeam -ProjectName $newProjectName).Id
 
       It 'Get-VSTeam ById Should return Teams' {
-         Get-VSTeam -ProjectName 'TeamModuleIntegration1' -Id $id | Should Not Be $null
+         Get-VSTeam -ProjectName $newProjectName -Id $global:id | Should Not Be $null
       }
 
       It 'Get-VSTeamMembers Should return TeamMembers' {
-         Get-VSTeamMember -ProjectName 'TeamModuleIntegration1' -Id $id | Should Not Be $null
+         Get-VSTeamMember -ProjectName $newProjectName -Id $global:id | Should Not Be $null
       }
 
       It 'Add-VSTeam should add a team' {
-         Add-VSTeam -ProjectName 'TeamModuleIntegration1' -Name 'testing' | Should Not Be $null
-         (Get-VSTeam -ProjectName 'TeamModuleIntegration1').Count | Should Be 2
+         Add-VSTeam -ProjectName $newProjectName -Name 'testing' | Should Not Be $null
+         (Get-VSTeam -ProjectName $newProjectName).Count | Should Be 2
       }
 
       It 'Update-VSTeam should update a team' {
-         Update-VSTeam -ProjectName 'TeamModuleIntegration1' -Name 'testing' -NewTeamName 'testing123'
-         Get-VSTeam -ProjectName 'TeamModuleIntegration1' -Name 'testing123' | Should Not Be $null
+         Update-VSTeam -ProjectName $newProjectName -Name 'testing' -NewTeamName 'testing123'
+         Get-VSTeam -ProjectName $newProjectName -Name 'testing123' | Should Not Be $null
       }
 
       It 'Remove-VSTeam should delete the team' {
-         Remove-VSTeam -ProjectName 'TeamModuleIntegration1' -Name 'testing123' -Force
-         Get-VSTeam -ProjectName 'TeamModuleIntegration1' | Where-Object { $_.Name -eq 'testing123'} | Should Be $null
+         Remove-VSTeam -ProjectName $newProjectName -Name 'testing123' -Force
+         Get-VSTeam -ProjectName $newProjectName | Where-Object { $_.Name -eq 'testing123'} | Should Be $null
       }
    }
 
@@ -222,7 +262,7 @@ Describe 'VSTeam Integration Tests' {
       }
 
       It 'Clear-VSTeamDefaultProject should clear project' {
-         $Global:PSDefaultParameterValues['*:projectName'] = 'TeamModuleIntegration1'
+         $Global:PSDefaultParameterValues['*:projectName'] = $newProjectName
          
          Clear-VSTeamDefaultProject
          
@@ -235,6 +275,15 @@ Describe 'VSTeam Integration Tests' {
    }
 
    AfterAll {
+      # Put everything back
+      Set-Location $originalLocation
+ 
+      if ($oAcct) { 
+         Add-VSTeamProfile -Account $oAcct -PersonalAccessToken $pat -Version $oVersion -Name $oName
+      }
+
+      Add-VSTeamAccount -Profile $oName
+
       Get-VSTeamProject | Remove-VSTeamProject -Force
 
       Remove-TeamAccount
